requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.3
CacheControl==0.12.5
cachetools==2.1.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==1.11
django-ckeditor==5.6.1
django-cors-headers==2.4.0
django-js-asset==1.1.0
django-multiselectfield==0.1.8
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
firebase-admin==2.13.0
google-api-core==1.3.0
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-firestore==0.29.0
google-cloud-storage==1.11.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.14.2
gunicorn==19.7.1
idna==2.7
instamojo-wrapper==1.1
jdcal==1.4
msgpack==0.5.6
mysqlclient==1.3.13
openpyxl==2.5.5
Pillow==5.2.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyJWT==1.6.4
PyPDF2==1.26.0
python-http-client==3.1.0
pytz==2018.4
qrcode==6.0
raven==6.9.0
reportlab==3.5.6
requests==2.19.1
rsa==3.4.2
sendgrid==5.4.1
six==1.11.0
urllib3==1.23
XlsxWriter==1.1.0
